+++ components/server/src/model/queries.py
@@ -1,6 +1,6 @@
 """Model queries."""
 
-from typing import cast, Any, List, Optional, Tuple
+from typing import cast, Any, Dict, List, Optional, Tuple
 
 from server_utilities.type import MetricId, ReportId, SourceId, SubjectId
 
@@ -33,3 +33,16 @@ def is_password_parameter(data_model, source_type: str, parameter: str) -> bool:
     # err on the safe side and assume it was a password type
     parameter_type = data_model["sources"][source_type]["parameters"].get(parameter, dict(type="password"))["type"]
     return str(parameter_type) == "password"
+
+
+def get_measured_attribute(data_model, metric_type: str, source_type: str) -> Optional[str]:
+    """Return the attribute of the entities of a source that are measured in the context of a metric. For example,
+    when using Jira as source for user story points, the points of user stories (the source entities) are summed to
+    arrive at the total number of user story points."""
+    return data_model["sources"][source_type]["entities"].get(metric_type, {}).get("measured_attribute")
+
+
+def get_attribute_type(entity: Dict[str, Dict[str, str]], attribute_key: Optional[str]) -> str:
+    """Look up the type of an entity attribute."""
+    return [attribute for attribute in entity["attributes"] if attribute["key"] == attribute_key][0].get(
+        "type", "text") if entity and attribute_key else "text"