+++ components/server/tests/routes/test_subject.py
@@ -6,7 +6,7 @@ from unittest.mock import Mock, patch
 from routes.subject import (
     delete_subject, post_move_subject, post_new_subject, post_subject_attribute, post_subject_copy_v3)
 
-from .fixtures import REPORT_ID, REPORT_ID2, SUBJECT_ID, SUBJECT_ID2, create_report
+from ..fixtures import REPORT_ID, REPORT_ID2, SUBJECT_ID, SUBJECT_ID2, create_report
 
 
 @patch("bottle.request")